feature-issue #12: Relocated communnication to partners module
